Move all Functional WS tests to one file Run all Functional WS tests synchronously to avoid connecting to peers instantiated for test only
